Create seperate files for data_extraction, responses and the database
