+++ .gitignore
@@ -1,7 +1,5 @@
 .venv/
 .vscode/
 __pycache__/
-*.py
-!main.py
 .env
 *.txt
