+++ requirements.txt
@@ -1,4 +1,5 @@
 absl-py==0.12.0
+astroid==2.6.2
 astunparse==1.6.3
 cachetools==4.2.2
 certifi==2020.12.5
@@ -16,18 +17,22 @@ grpcio==1.34.1
 gunicorn==20.1.0
 h5py==3.1.0
 idna==2.10
+isort==5.9.2
 itsdangerous==2.0.1
 Jinja2==3.0.1
 keras-nightly==2.5.0.dev2021032900
 Keras-Preprocessing==1.1.2
+lazy-object-proxy==1.6.0
 Markdown==3.3.4
 MarkupSafe==2.0.1
+mccabe==0.6.1
 numpy==1.19.5
 oauthlib==3.1.1
 opt-einsum==3.3.0
 protobuf==3.17.3
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
+pylint==2.9.3
 pymongo==3.11.4
 python-dotenv==0.17.1
 requests==2.25.1
@@ -40,6 +45,7 @@ tensorboard-plugin-wit==1.8.0
 tensorflow==2.5.0
 tensorflow-estimator==2.5.0
 termcolor==1.1.0
+toml==0.10.2
 typing-extensions==3.7.4.3
 urllib3==1.26.5
 Werkzeug==2.0.1
